fix #3179: creating a common base for the default kube and os clients
